+++ backend/src/main/java/org/example/backend/users/domain/dto/admin/AdminResDto.java
@@ -1,15 +1,12 @@
-package org.example.backend.users.domain.dto;
+package org.example.backend.users.domain.dto.admin;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class AdminReqDto {
+public class AdminResDto {
 
     private String loginId;
-
-    private String password;
-
     private String email;
 }
