+++ backend/src/main/java/org/example/backend/users/domain/entity/Role.java
@@ -1,5 +1,5 @@
 package org.example.backend.users.domain.entity;
 
 public enum Role {
-    ADMIN, USER
+    ADMIN, MEMBER
 }
