+++ backend/src/main/java/org/example/backend/users/domain/entity/Users.java
@@ -10,7 +10,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.backend.users.domain.dto.SignInReqDto;
+import org.example.backend.users.domain.dto.admin.SignInReqDto;
 
 @Entity
 @Getter
